+++ changelog.md
@@ -1,4 +1,12 @@
 # Changelog
 
-**./test-parts/test-possible-routes-valid.js - handleCount**
-* Renamed "Infinite Routes" test to "Infinite"
+**./test-parts/test-possible-routes-valid.js**
+* Created "Arguments" test group. (handleArguments)
+* See below entry for contents.
+
+---
+
+**./test-parts/test-possible-routes-invalid.js**
+* Moved the following tests to 'test-possible-routes-valid.js' under "Arguments"
+	* "Arguments" - "Empty Criteria"
+	* "Arguments" - "Impossible Route"